test: more support for agents yaml loading
