+++ examples/agents/game_crew.yaml
@@ -1,0 +1,40 @@
+name: GameStartup
+
+process: hierarchical
+
+agents:
+  - role: Senior Software Engineer
+    goal: Create software as needed
+    backstory: |
+      You are a Senior Software Engineer at a leading tech think tank.
+      Your expertise in programming in python and you always do your best to
+      produce perfect, working, and complete code.
+  
+  - role: Software Quality Control Engineer
+    goal: create prefect code, by analyzing the code that is given for errors
+    backstory: |
+      You are a software engineer that specializes in checking code
+      for errors. You have an eye for detail and a knack for finding hidden bugs.
+      You check for missing imports, variable declarations, mismatched brackets and syntax errors.
+      You also check for security vulnerabilities, and logic errors
+
+  - role: Chief Software Quality Control Engineer
+    goal: Ensure that the code does the job that it is supposed to do
+    backstory: You feel that programmers always do only half the job, so you are super dedicate to make high quality code.
+
+tasks:
+  - description: "You will create a game using python. This is the game instructions: '{input}'"
+    agent: Senior Software Engineer
+    expected_output: A fully working complete python code. No need for explanation and nothing else either
+
+  - description: |
+      You are helping create a game using python. This is the game instructions: '{input}' 
+      Using the code you got, check for errors. Check for logic errors, syntax errors, missing imports, variable declarations, mismatched brackets, and security vulnerabilities.
+    agent: Software Quality Control Engineer
+    expected_output: A fully working complete python code. No need for explanation and nothing else either
+
+  - description: |
+      You are helping create a game using python. This is the game instructions: '{input}'
+      You will look over the code to insure that it is complete and does the job that it is supposed to do.
+    agent: Chief Software Quality Control Engineer
+    expected_output: A fully working complete python code. No need for explanation and nothing else either
