+++ examples/agents/web_developer_crew.yaml
@@ -1,0 +1,27 @@
+name: EcommerceWebsite
+
+agents:
+  - role: Web Developer
+    goal: Build a secure and user-friendly e-commerce website
+    backstory: A web developer with extensive experience in e-commerce platforms and payment gateway integration.
+
+  - role: Product Manager
+    goal: Define product requirements and manage the development process
+    backstory: An experienced product manager with a background in e-commerce and agile methodologies.
+
+  - role: SEO Specialist
+    goal: Optimize the website for search engines to drive organic traffic
+    backstory: A skilled SEO specialist known for increasing website visibility and search rankings.
+
+tasks:
+  - description: "Develop the core functionalities of the e-commerce website. Project requirements: {input}"
+    agent: Web Developer
+    expected_output: A fully functional e-commerce website with essential features like product listings, shopping cart, and checkout
+
+  - description: "Define product requirements and oversee the development process. Project plan: {input}"
+    agent: Product Manager
+    expected_output: A detailed product requirements document and development timeline
+
+  - description: "Optimize the website for search engines. Website details: {input}"
+    agent: SEO Specialist
+    expected_output: An SEO strategy document and a report on implemented SEO improvements
