+++ layout.yaml
@@ -1,0 +1,14 @@
+## Sage Agents Layout
+
+name: GameStartup # The name of the crew
+
+agents: # a list of agents for the crew
+ - role: # role of the agent
+   goal: # goal of the agent
+   backstory: # backstory
+   
+
+tasks: # a list of tasks
+ - desciption: # the task description that will be assigned to the agent
+   agent: # the agent to used
+   expected_output: # description of the agent's expected output
