+++ sage/validators/starters.py
@@ -5,8 +5,10 @@ from chainlit import Starter
 from pydantic import (
     BaseModel,
     Field,
+    ValidationError,
     model_validator,
 )
+import yaml
 from sage.utils.exceptions import ConfigException
 
 
@@ -31,3 +33,26 @@ class Starters(BaseModel):
     """Starters config model"""
 
     starters: List[StarterConfig] = Field(default=[])
+
+
+def load_and_validate_starters_yaml(file_path: str | None) -> Starters:
+    """Validates the chat starters yaml is valid"""
+    if file_path is None:
+        return []
+    try:
+        with open(file_path, "r") as file:
+            data = yaml.safe_load(file)
+            if data is None:
+                raise ConfigException("Starters content cannot be empty")
+    except FileNotFoundError:
+        raise ConfigException(f"Starters file not found at path: {file_path}")
+    except yaml.YAMLError as exc:
+        raise ConfigException(f"Error parsing YAML: {exc}")
+
+    # Validate data with pydantic
+    try:
+        starters_config = Starters(**data).starters
+    except ValidationError as ve:
+        raise ConfigException(f"Validation error in starters YAML: {ve}")
+
+    return starters_config
